pyproject.toml
[tool.poetry]
name = "py-map-generation"
version = "0.1.0"
description = ""
authors = ["Fluffy <michal-gdula@protonmail.com>"]
readme = "README.md"
packages = [{include = "py_map_generation"}]

[tool.poetry.dependencies]
python = "^3.11"
numpy = "^1.25.2"
pillow = "^10.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
